-Simplified names for COM models of Alarm and Audit services -Renamed model of COM according to Basic Services
