+++ MetasysServicesCom/Interfaces/IComVariantMultiple.cs
@@ -8,7 +8,7 @@ namespace JohnsonControls.Metasys.ComServices
     /// A structure for grouping Variant values with the same id.
     /// </summary>
     [ComVisible(true)]
-    [Guid("3b7dbcd6-f507-4cbe-bd0c-ecd418bed227")]
+    [Guid("b5e6f102-c95c-484c-857f-6507c604c27e")]
     [InterfaceType(ComInterfaceType.InterfaceIsIDispatch)]
     public interface IComVariantMultiple
     {
@@ -16,6 +16,6 @@ namespace JohnsonControls.Metasys.ComServices
         string Id { set; get; }
 
         /// <summary>The list of Variants. </summary>
-        object Variants { set; get; } // Note: need a generic object as return type in order to map correctly to VBA type array (can't assign to array error)
+        object Values { set; get; } // Note: need a generic object as return type in order to map correctly to VBA type array (can't assign to array error)
     }
 }
