+++ MetasysServicesCom/Models/ComVariantMultiple.cs
@@ -7,7 +7,7 @@ namespace JohnsonControls.Metasys.ComServices
     /// <summary>
     /// A specialized COM structure for grouping Variant values with the same id.
     /// </summary>    
-    [Guid("57946469-9f3a-46f7-a7d8-6c38a406ed4f")]
+    [Guid("4e6ce75b-8495-442a-ad97-85721e2ab876")]
     [ComVisible(true)]
     [ClassInterface(ClassInterfaceType.None)]   
     public class ComVariantMultiple:IComVariantMultiple
@@ -18,6 +18,6 @@ namespace JohnsonControls.Metasys.ComServices
         public string Id { set; get; }
 
         /// <inheritdoc/>
-        public object Variants { set; get; }
+        public object Values { set; get; }
     }
 }
